Set system time from pump when it’s way off
The application will attempt to update the system time from ntp, but
that's not always available and some embedded systems like the Raspberry
Pi don't have a battery-backed realtime clock. In the case where the
time is set to the epoch, the application will fall back to querying the
pump date/time.