Move aps loop code into separate project within the poncho
There is a circular dependency between `ui` and `fw`, since `fw` needs
to package `ui` and `ui` needs access to display loop code contained in
`fw`. As a result, `aps` has been created as a dependency of both. The
goal is to be able to run `ui` standalone on the host for hot code
reload and faster feedback cycles than running inside `fw` in host mode.